adaptacia wpisow do material3 i zmiana motyru aplikaci

+++ app/src/main/java/pl/kalisz/ak/rafal/peczek/mojepomiary/pomiary/PomiaryEdytuj.java
@@ -1,16 +1,22 @@
 package pl.kalisz.ak.rafal.peczek.mojepomiary.pomiary;
 
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
+import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.material.dialog.MaterialAlertDialogBuilder;
+import com.google.android.material.textfield.TextInputLayout;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -19,6 +25,7 @@ import java.util.List;
 import pl.kalisz.ak.rafal.peczek.mojepomiary.R;
 import pl.kalisz.ak.rafal.peczek.mojepomiary.entity.Jednostka;
 import pl.kalisz.ak.rafal.peczek.mojepomiary.entity.Pomiar;
+import pl.kalisz.ak.rafal.peczek.mojepomiary.jednostki.JednostkiEdytuj;
 import pl.kalisz.ak.rafal.peczek.mojepomiary.repository.UsersRoomDatabase;
 
 public class PomiaryEdytuj extends AppCompatActivity {
@@ -27,43 +34,57 @@ public class PomiaryEdytuj extends AppCompatActivity {
     private int pomiarId;
 
     Pomiar pomiar;
-    private EditText nazwa, notatka;
-    private Spinner jednostki;
+    private TextInputLayout nazwa, notatka;
+    private AutoCompleteTextView jednostki;
+    private TextInputLayout jednostkiL;
     private UsersRoomDatabase database;
     private List<Jednostka> listaJednostek;
     private int idWybranejJednostki;
+    private String textWybranejJednostki;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_pomiary_edytuj);
 
+        idWybranejJednostki = 0;
+        textWybranejJednostki = "";
         pomiarId = (Integer) getIntent().getExtras().get(EXTRA_Pomiar_ID);
 
-        nazwa = (EditText) findViewById(R.id.editTextNazwa);
-        notatka = (EditText) findViewById(R.id.editTextJednostka);
-        jednostki = (Spinner) findViewById(R.id.spinner);
+        nazwa = (TextInputLayout) findViewById(R.id.editTextNazwaLayout);
+        notatka = (TextInputLayout) findViewById(R.id.editTextJednostkaLayout);
+        jednostki = (AutoCompleteTextView) findViewById(R.id.spinner);
+        jednostkiL = (TextInputLayout) findViewById(R.id.spinnerLayout);
 
         database = UsersRoomDatabase.getInstance(getApplicationContext());
         listaJednostek = database.localJednostkaDao().getAll();
         pomiar = database.localPomiarDao().findById(pomiarId);
-        nazwa.setText(pomiar.getNazwa());
-        notatka.setText(pomiar.getNotatka());
+        nazwa.getEditText().setText(pomiar.getNazwa());
+        notatka.getEditText().setText(pomiar.getNotatka());
 
         //jednostka.
         ArrayList<String> data = new ArrayList<>();
-        data.add("Wybierz jednostkę");
         for (Jednostka jednostka: listaJednostek){
-            if(pomiar.getIdJednostki() == jednostka.getId())
+            String element = jednostka.getNazwa()+" "+jednostka.getWartosc();
+            if(pomiar.getIdJednostki() == jednostka.getId()) {
                 idWybranejJednostki = data.size();
-            data.add(jednostka.getNazwa()+" "+jednostka.getWartosc());
+                textWybranejJednostki = element;
+            }
+            data.add(element);
         }
 
         ArrayAdapter adapter = new ArrayAdapter ( this, android.R.layout.simple_spinner_dropdown_item, data);
-
         jednostki.setAdapter(adapter);
-        jednostki.setSelection(idWybranejJednostki);
-        jednostki.setEnabled(false);
+
+        jednostki.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+                idWybranejJednostki = position;
+            }
+        });
+
+        jednostki.setText(textWybranejJednostki, false);
+        jednostkiL.setEnabled(false);
     }
 
     @Override
@@ -84,15 +105,26 @@ public class PomiaryEdytuj extends AppCompatActivity {
                 anuluj.setEnabled(true);
                 nazwa.setEnabled(true);
                 notatka.setEnabled(true);
-                jednostki.setEnabled(true);
+                jednostkiL.setEnabled(true);
                 return true;
-            case R.id.drop:
-                if (database != null) {
-                    Pomiar pomiar = database.localPomiarDao().findById(pomiarId);
-                    database.localPomiarDao().delete(pomiar);
-                }
-                finish();
+            case R.id.drop:{
+                MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(PomiaryEdytuj.this);
+                builder.setMessage("Czy na pewno usunąć");
+                builder.setCancelable(false);
+                builder.setPositiveButton("Tak", (DialogInterface.OnClickListener) (dialog, which) -> {
+                    if (database != null) {
+                        Pomiar pomiar = database.localPomiarDao().findById(pomiarId);
+                        database.localPomiarDao().delete(pomiar);
+                    }
+                    finish();
+                });
+
+                builder.setNegativeButton("Nie", (DialogInterface.OnClickListener) (dialog, which) -> {
+                    dialog.cancel();
+                });
+                builder.show();
                 return true;
+            }
             default:
                 return super.onOptionsItemSelected(item);
         }
@@ -106,28 +138,27 @@ public class PomiaryEdytuj extends AppCompatActivity {
         anuluj.setEnabled(false);
         nazwa.setEnabled(false);
         notatka.setEnabled(false);
-        jednostki.setEnabled(false);
-        nazwa.setText(pomiar.getNazwa());
-        notatka.setText(pomiar.getNotatka());
-        jednostki.setSelection(idWybranejJednostki);
+        jednostkiL.setEnabled(false);
+        nazwa.getEditText().setText(pomiar.getNazwa());
+        notatka.getEditText().setText(pomiar.getNotatka());
+        jednostki.setText(textWybranejJednostki, false);
     }
 
     public void aktualizuj(View view) {
-        String nazwa = this.nazwa.getText().toString();
-        String notatka = this.notatka.getText().toString();
-        int jednostka = (int) this.jednostki.getSelectedItemId();
+        String nazwa = this.nazwa.getEditText().getText().toString();
+        String notatka = this.notatka.getEditText().getText().toString();
 
-        if(jednostka != 0 && nazwa.length()>=2 && notatka.length()>=2) {
+        if( jednostki.getText().length()>0 && nazwa.length()>=2 && notatka.length()>=2) {
             nazwa = nazwa.substring(0, 1).toUpperCase() + nazwa.substring(1).toLowerCase();
             notatka = notatka.substring(0, 1).toUpperCase() + notatka.substring(1);
             if (notatka.charAt(notatka.length() - 1) != '.')
                 notatka += '.';
-            jednostka = listaJednostek.get(jednostka - 1).getId();
+            int jednostkaId = listaJednostek.get(idWybranejJednostki).getId();
 
             Pomiar pomiar = database.localPomiarDao().findById(pomiarId);
             pomiar.setNazwa(nazwa);
             pomiar.setNotatka(notatka);
-            pomiar.setIdJednostki(jednostka);
+            pomiar.setIdJednostki(jednostkaId);
             pomiar.setDataAktualizacji(new Date());
 
             database.localPomiarDao().insert(pomiar);
